+++ jarust_core/Cargo.toml
@@ -23,7 +23,6 @@ tokio = { workspace = true, features = ["sync", "time"] }
 tracing.workspace = true
 
 [features]
-default = ["tokio-rt", "use-native-tls"]
 tokio-rt = ["jarust_rt/tokio-rt", "jarust_interface/tokio-rt"]
 use-native-tls = ["jarust_interface/use-native-tls"]
 use-rustls = ["jarust_interface/use-rustls"]
