+++ jarust_interface/Cargo.toml
@@ -29,15 +29,14 @@ tracing.workspace = true
 uuid = { version = "1.11.0", features = ["fast-rng", "v4"] }
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
-rustls = { version = "0.23.13", optional = true }
-rustls-native-certs = { version = "0.8.0", optional = true }
+rustls = { version = "0.23.20", optional = true }
+rustls-native-certs = { version = "0.8.1", optional = true }
 tokio-tungstenite = "0.26.1"
 
 [target.'cfg(target_family = "wasm")'.dependencies]
 getrandom = { version = "0.2.12", features = ["js"] }
 
 [features]
-default = ["use-native-tls", "tokio-rt"]
 use-native-tls = ["tokio-tungstenite/native-tls"]
 use-rustls = ["rustls", "rustls-native-certs", "tokio-tungstenite/__rustls-tls"]
 tokio-rt = ["jarust_rt/tokio-rt"]
